refactor(measurement): split sweep execution and views
